requirements.txt
numpy
pandas
torch
torchvision
Pillow
torchtext
scikit-learn
